# typescript/wind-srv/openapi.yaml
openapi: 3.0.3
info:
  title: wind-srv — Wind Workflow Schema API
  version: 1.0.7bdbe3b1
  description: 'REST API for the wind workflow schema: offices, titles, tasks, workflow graphs, runtime instances, tickets,
    and receipts.


    Endpoint inventory generated from source route registrations by `nexus/tools/api-docs/`. Request/response bodies are generic;
    field-level contracts live in the service code and its README.'
servers:
- url: http://localhost:3300
...
paths:
...
  /api/instances:
    get:
      tags:
      - instances
      operationId: get_instances
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonBody'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      summary: List instances (optionally filter by status or workflow_id)
    post:
      tags:
      - instances
      operationId: post_instances
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonBody'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      summary: Start a workflow instance Creates an instance and tickets for the entrypoint node(s)
  /api/instances/{id}:
    get:
      tags:
      - instances
      operationId: get_instances_id
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonBody'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      summary: Get instance by ID (with tickets)
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /api/instances/{id}/advance:
    post:
      tags:
      - instances
      operationId: post_instances_id_advance
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonBody'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      summary: Advance — complete a ticket with an outcome, create next tickets This is the core workflow traversal step
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /api/instances/{id}/execute:
    post:
      tags:
      - instances
      operationId: post_instances_id_execute
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonBody'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      summary: Execute — run the harness for a ticket's task
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /api/instances/{id}/pause:
    post:
      tags:
      - instances
      operationId: post_instances_id_pause
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonBody'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      summary: Pause an instance
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /api/instances/{id}/resume:
    post:
      tags:
      - instances
      operationId: post_instances_id_resume
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonBody'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      summary: Resume a paused instance
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /api/instances/{id}/run:
    post:
      tags:
      - instances
      operationId: post_instances_id_run
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonBody'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      summary: 'Run an instance to completion (loop: execute → advance → … until terminal).'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /api/instances/{id}/stop:
    post:
      tags:
      - instances
      operationId: post_instances_id_stop
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonBody'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      summary: Stop (cancel) an instance
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
...
components:
  schemas:
    JsonBody:
      type: object
      additionalProperties: true
      description: Generic JSON body (fields are service-specific).
    Error:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code (e.g. not_found, validation_failed).
        message:
          type: string
          description: Human-readable error detail.
      additionalProperties: true
...